Common dataset class (#115)
* first spilt of stats functions

* finishing stats computation isolation

* lint

* create dataAccesor class and implem

* factoring out attributes from titan and poesy datasets

* separating titan_cli from titan dataset file

* adding poesy cli and spearating from main dataset

* lint

* named tensors isolated + rationalizing imports

* lint again

* fixing imports

* reformat

* lint2

* corrections

* poesy ok with hardcoded stats name

* titan small changes

* correction get_dataset_path

* corrections

* factorize sample list

* fix imports

* remove named_tensors from py4cast (now mfai)

* named tensor from mfai

* factorize get_param_tensor

* fix poesy tensor shape

* moving concrete datasets to cli only

* lint

* fix Titan, Dummy and lint

* fix imports and dataset cli

* fix imports

* fix import Titan

* typo failing test models

* create dummy config

* lint

* fix dummy dataset for tests + setup class methods

* fix imports

* fix dir for dummy config

* lint

* fix config file

* fix dummy config file name

* simplify param list with accessor

* removing useless imports

* flexibilised dataset name

* lint

* remove registry dependency on train

* Update add_features_contribute.md

* Update add_features_contribute.md

* modify parameter naming, fix bugs in input-output distrbution. Working with arpege-lbc

* lint

* Update py4cast/datasets/base.py


* modifying titan lbc to compat with refacto

* add doc to exists function

* lint

* lint and add suggestion

* Update py4cast/datasets/access.py


* Update py4cast/datasets/access.py


* fix naming of aggregate /stat_name

* remove notImplementedError from decorated abstract methods

---------
