+++ py4cast/settings.py
@@ -3,3 +3,4 @@ from pathlib import Path
 
 ROOTDIR = Path(os.environ.get("PY4CAST_ROOTDIR", "/scratch/shared/py4cast"))
 CACHE_DIR = ROOTDIR / "cache"
+DEFAULT_CONFIG_DIR = Path(__file__).parents[1] / "config"
